add R and SLURM scripts to run on ACCRE to close #190

+++ 01-slurm-experiments.R
@@ -7,19 +7,25 @@
 ##
 #######
 
+args <- commandArgs(trailingOnly = TRUE)
+
 #expected command line parameters
-n_boot = 2
-task_ID = 1
+n_boot = as.numeric(args[[1]])
+task_ID = as.numeric(args[[2]])
+
+#compatibility with SLURM
+.libPaths("~/R/x86_64-pc-linux-gnu-library/3.6")
+chooseCRANmirror(graphics=FALSE, ind=1)
+
+# Get required packages
+if (!require("pacman")) install.packages("pacman")
+pacman::p_load(knitr, stringr, readr, fs)
 
 #other parameters
 notebook_file = '81-compute-suicidal'
 outcome = 'h5mn8'
-results_directory = str_c(outcome, '/results_run_1')
+results_directory = str_c(outcome, '/results_run_2')
 kfold_file = str_c(outcome, '/kfold_assign_3895_tts_9384.csv')
-
-# Get required packages
-if (!require("pacman")) install.packages("pacman")
-pacman::p_load(knitr, readr, fs)
 
 # Convert the 70 series notebook to a script
 purl(paste0(notebook_file, '.Rmd'))
@@ -31,5 +37,7 @@ r_file = paste0(notebook_file, '.R')
 source(r_file)
 
 # Delete the generated file
-file_delete(r_file)
+if(file_exists(r_file)){
+  file_delete(r_file) 
+}
